Change the type of desired number of rows

+++ be/src/exec/assert_num_rows_node.h
@@ -34,7 +34,7 @@ public:
     virtual Status close(RuntimeState* state);
 
 private:
-    int _desired_num_rows;
+    int64_t _desired_num_rows;
     const std::string _subquery_string;
 };
